# Leverage `xsync.Map` instead of custom sync map implementation