--- a/go.mod
+++ b/go.mod
@@ -1,15 +1,12 @@
 module github.com/linkedin/diderot
 
-go 1.23.0
-
-// This is temporary, will be replaced with the new major release
-replace github.com/puzpuzpuz/xsync/v3 => github.com/PapaCharlie/xsync/v3 v3.0.0-20250326183228-c5848d82e85f
+go 1.24.1
 
 require (
 	github.com/envoyproxy/go-control-plane v0.13.4
 	github.com/envoyproxy/go-control-plane/envoy v1.32.4
 	github.com/google/go-cmp v0.6.0
-	github.com/puzpuzpuz/xsync/v3 v3.5.1
+	github.com/puzpuzpuz/xsync/v4 v4.0.0
 	github.com/stretchr/testify v1.10.0
 	golang.org/x/time v0.5.0
 	google.golang.org/genproto/googleapis/rpc v0.0.0-20241202173237-19429a94021a