--- a/internal/cache/resource_map.go
+++ b/internal/cache/resource_map.go
@@ -17,41 +17,45 @@ func NewResourceMap[K comparable, V any]() *ResourceMap[K, V] {
 }
 
 // Compute first creates the value for the given key using the given function if no corresponding
-// entry exists. Then it executes the given compute function
+// entry exists, then it executes the given compute function. It returns the value itself, and a
+// boolean indicating whether the value was created.
 func (m *ResourceMap[K, V]) Compute(
 	key K,
 	newValue func(key K) V,
 	compute func(value V),
-) V {
-	v, _ := (*xsync.MapOf[K, V])(m).ComputeV2(key, func(v V, loaded bool) (_ V, op xsync.ComputeOp) {
+) (v V, created bool) {
+	v, _ = (*xsync.MapOf[K, V])(m).ComputeV2(key, func(v V, loaded bool) (_ V, op xsync.ComputeOp) {
 		if !loaded {
 			v = newValue(key)
 			op = xsync.Update
+			created = true
 		}
 		compute(v)
 		return v, op
 	})
-	return v
+	return v, created
 }
 
 // ComputeIfPresent invokes the given function only if a corresponding entry exists in the map for
 // the given key.
-func (m *ResourceMap[K, V]) ComputeIfPresent(key K, f func(value V)) {
+func (m *ResourceMap[K, V]) ComputeIfPresent(key K, f func(value V)) (wasPresent bool) {
 	(*xsync.MapOf[K, V])(m).ComputeV2(key, func(oldValue V, loaded bool) (_ V, op xsync.ComputeOp) {
 		if !loaded {
 			return oldValue, op
 		}
 
 		f(oldValue)
+		wasPresent = true
 		return oldValue, op
 	})
+	return wasPresent
 }
 
 // ComputeDeletion loads the entry from the map if it still exists, then executes the given condition
 // function with the value. If the condition returns true, the entry is deleted from the map,
 // otherwise nothing happens. As a "compute" function, the condition is executed synchronously, in
 // other words, it is guaranteed that no other "compute" functions are executing on that entry.
-func (m *ResourceMap[K, V]) ComputeDeletion(key K, condition func(value V) (deleteEntry bool)) {
+func (m *ResourceMap[K, V]) ComputeDeletion(key K, condition func(value V) (deleteEntry bool)) (deleted bool) {
 	(*xsync.MapOf[K, V])(m).ComputeV2(key, func(oldValue V, loaded bool) (_ V, op xsync.ComputeOp) {
 		if !loaded {
 			return oldValue, op
@@ -59,10 +63,12 @@ func (m *ResourceMap[K, V]) ComputeDeletion(key K, condition func(value V) (dele
 
 		if condition(oldValue) {
 			op = xsync.Delete
+			deleted = true
 		}
-
 		return oldValue, op
 	})
+
+	return deleted
 }
 
 // Range returns an [iter.Seq2] that will iterate over all entries in this map.