The `ResourceMap` was an attempt to make the stdlib `sync.Map` have `compute` methods like Java's `ConcurrentHashMap`. However, this ended up being far too complex and error prone. Instead of reinventing the wheel, leveraging the sync map implementation provided by the `github.com/puzpuzpuz/xsync` makes things significantly cleaner. Functionally, there should be no behavioral changes, but new features should be less likely to cause problems down the line.

Additionally, this optimizes the notification loop pretty significantly. In short, the old implementation captured the `SubscriberSetIterator` on-heap, preventing the compiler from inlining it, causing heap allocations every loop and dragging performance down with it. This relatively simple change is functionally identical and just as readable, except that it all gets inlined :)

Tested by running all the unit tests with `-count=100` overnight, to double check that there are no known race conditions.

This is an apples to apples comparison of before and after this change, and we can see that there are no significant slowdowns whatsoever. In fact, this allocates less memory and therefore in the long term will be better. So this code is simpler, performs the same and is less error prone, so it's a win win!
```
goos: linux
goarch: amd64
pkg: github.com/linkedin/diderot
cpu: AMD EPYC 7763 64-Core Processor
                                                  │     old      │                  new                  │
                                                  │    sec/op    │    sec/op     vs base                 │
CacheThroughput/____1_subscribers/____1_entries-8   49.70n ±  2%   56.51n ±  3%   +13.72% (p=0.000 n=10)
CacheThroughput/____1_subscribers/___10_entries-8   159.8n ±  2%   334.8n ±  1%  +109.48% (p=0.000 n=10)
CacheThroughput/____1_subscribers/__100_entries-8   1.251µ ±  9%   1.631µ ±  2%   +30.47% (p=0.000 n=10)
CacheThroughput/____1_subscribers/_1000_entries-8   11.69µ ±  1%   11.99µ ±  1%    +2.58% (p=0.000 n=10)
CacheThroughput/___10_subscribers/____1_entries-8   87.22n ± 11%   56.99n ±  2%   -34.65% (p=0.000 n=10)
CacheThroughput/___10_subscribers/___10_entries-8   165.1n ±  3%   240.7n ±  3%   +45.86% (p=0.000 n=10)
CacheThroughput/___10_subscribers/__100_entries-8   1.276µ ± 18%   1.903µ ±  6%   +49.20% (p=0.000 n=10)
CacheThroughput/___10_subscribers/_1000_entries-8   11.71µ ±  1%   11.96µ ±  1%    +2.10% (p=0.000 n=10)
CacheThroughput/__100_subscribers/____1_entries-8   194.3n ±  3%   252.0n ±  3%   +29.66% (p=0.000 n=10)
CacheThroughput/__100_subscribers/___10_entries-8   173.1n ±  5%   237.3n ±  6%   +37.13% (p=0.000 n=10)
CacheThroughput/__100_subscribers/__100_entries-8   1.244µ ±  5%   1.745µ ±  7%   +40.29% (p=0.000 n=10)
CacheThroughput/__100_subscribers/_1000_entries-8   11.81µ ±  1%   12.09µ ±  1%    +2.44% (p=0.000 n=10)
CacheThroughput/__100_subscribers/10000_entries-8   122.3µ ±  0%   123.8µ ±  8%    +1.21% (p=0.000 n=10)
CacheThroughput/_1000_subscribers/____1_entries-8   192.4n ±  2%   272.9n ±  5%   +41.78% (p=0.000 n=10)
CacheThroughput/_1000_subscribers/___10_entries-8   169.2n ±  3%   951.2n ± 12%  +462.04% (p=0.000 n=10)
CacheThroughput/_1000_subscribers/__100_entries-8   1.354µ ± 11%   2.427µ ± 12%   +79.28% (p=0.000 n=10)
CacheThroughput/_1000_subscribers/_1000_entries-8   12.20µ ±  2%   12.48µ ±  1%    +2.27% (p=0.000 n=10)
CacheThroughput/10000_subscribers/__100_entries-8   1.636µ ±  9%   2.888µ ± 22%   +76.58% (p=0.000 n=10)
geomean                                             1.016µ         1.397µ         +37.51%

                                                  │       old       │                  new                   │
                                                  │      B/op       │    B/op      vs base                   │
CacheThroughput/____1_subscribers/____1_entries-8    0.000 ±   0%     0.000 ±  0%         ~ (p=1.000 n=10) ¹
CacheThroughput/____1_subscribers/___10_entries-8    0.000 ±   0%     0.000 ±  0%         ~ (p=1.000 n=10) ¹
CacheThroughput/____1_subscribers/__100_entries-8    0.000 ±   0%     1.500 ± 33%         ? (p=0.000 n=10)
CacheThroughput/____1_subscribers/_1000_entries-8    9.000 ±  11%     2.000 ±  0%   -77.78% (p=0.000 n=10)
CacheThroughput/___10_subscribers/____1_entries-8    11.50 ±  22%      0.00 ±  0%  -100.00% (p=0.000 n=10)
CacheThroughput/___10_subscribers/___10_entries-8    3.000 ±  33%     0.000 ±  0%  -100.00% (p=0.000 n=10)
CacheThroughput/___10_subscribers/__100_entries-8    4.500 ± 278%     2.000 ±  0%   -55.56% (p=0.002 n=10)
CacheThroughput/___10_subscribers/_1000_entries-8   22.000 ±  23%     2.000 ±  0%   -90.91% (p=0.000 n=10)
CacheThroughput/__100_subscribers/____1_entries-8    47.00 ±   4%      0.00 ±  0%  -100.00% (p=0.000 n=10)
CacheThroughput/__100_subscribers/___10_entries-8    7.500 ±  47%     0.000 ±  0%  -100.00% (p=0.000 n=10)
CacheThroughput/__100_subscribers/__100_entries-8    5.000 ±  40%     0.000 ±  0%  -100.00% (p=0.000 n=10)
CacheThroughput/__100_subscribers/_1000_entries-8   45.500 ±  32%     2.000 ±  0%   -95.60% (p=0.000 n=10)
CacheThroughput/__100_subscribers/10000_entries-8    854.5 ±   3%     284.0 ± 11%   -66.76% (p=0.000 n=10)
CacheThroughput/_1000_subscribers/____1_entries-8    46.00 ±   4%      0.00 ±  0%  -100.00% (p=0.000 n=10)
CacheThroughput/_1000_subscribers/___10_entries-8    6.000 ±  67%     0.000 ±  0%  -100.00% (p=0.000 n=10)
CacheThroughput/_1000_subscribers/__100_entries-8    15.00 ±  40%      0.00 ±  0%  -100.00% (p=0.000 n=10)
CacheThroughput/_1000_subscribers/_1000_entries-8   57.000 ±  23%     2.000 ±  0%   -96.49% (p=0.000 n=10)
CacheThroughput/10000_subscribers/__100_entries-8    32.00 ±  78%      0.00 ±  0%  -100.00% (p=0.000 n=10)
geomean                                                           ²                ?                       ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                                  │       old       │                   new                    │
                                                  │    allocs/op    │  allocs/op   vs base                     │
CacheThroughput/____1_subscribers/____1_entries-8    0.000 ±   0%     0.000 ±  0%         ~ (p=1.000 n=10) ¹
CacheThroughput/____1_subscribers/___10_entries-8    0.000 ±   0%     0.000 ±  0%         ~ (p=1.000 n=10) ¹
CacheThroughput/____1_subscribers/__100_entries-8    0.000 ±   0%     0.000 ±  0%         ~ (p=1.000 n=10) ¹
CacheThroughput/____1_subscribers/_1000_entries-8    0.000 ±   0%     0.000 ±  0%         ~ (p=1.000 n=10) ¹
CacheThroughput/___10_subscribers/____1_entries-8    0.000 ±   0%     0.000 ±  0%         ~ (p=1.000 n=10) ¹
CacheThroughput/___10_subscribers/___10_entries-8    0.000 ±   0%     0.000 ±  0%         ~ (p=1.000 n=10) ¹
CacheThroughput/___10_subscribers/__100_entries-8    0.000 ±   0%     0.000 ±  0%         ~ (p=1.000 n=10) ¹
CacheThroughput/___10_subscribers/_1000_entries-8    0.000 ±   0%     0.000 ±  0%         ~ (p=1.000 n=10)
CacheThroughput/__100_subscribers/____1_entries-8    1.000 ±   0%     0.000 ±  0%  -100.00% (p=0.000 n=10)
CacheThroughput/__100_subscribers/___10_entries-8    0.000 ±   0%     0.000 ±  0%         ~ (p=1.000 n=10) ¹
CacheThroughput/__100_subscribers/__100_entries-8    0.000 ±   0%     0.000 ±  0%         ~ (p=1.000 n=10) ¹
CacheThroughput/__100_subscribers/_1000_entries-8    1.000 ±   0%     0.000 ±  0%  -100.00% (p=0.000 n=10)
CacheThroughput/__100_subscribers/10000_entries-8   25.000 ±   4%     5.000 ± 20%   -80.00% (p=0.000 n=10)
CacheThroughput/_1000_subscribers/____1_entries-8    1.000 ±   0%     0.000 ±  0%  -100.00% (p=0.000 n=10)
CacheThroughput/_1000_subscribers/___10_entries-8    0.000 ±   0%     0.000 ±  0%         ~ (p=1.000 n=10) ¹
CacheThroughput/_1000_subscribers/__100_entries-8    0.000 ±   0%     0.000 ±  0%         ~ (p=1.000 n=10)
CacheThroughput/_1000_subscribers/_1000_entries-8    1.000 ± 100%     0.000 ±  0%  -100.00% (p=0.000 n=10)
CacheThroughput/10000_subscribers/__100_entries-8   500.0m ±    ?      0.0m ±  0%  -100.00% (p=0.033 n=10)
geomean                                                           ²                ?                       ² ³
¹ all samples are equal
² summaries must be >0 to compute geomean
³ ratios must be >0 to compute geomean
```